+++ src/main/java/com/jomkie/sbmp/service/impl/KongFuServiceImpl.java
@@ -11,7 +11,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 import java.util.stream.LongStream;
 
 /**
@@ -35,12 +34,12 @@ public class KongFuServiceImpl extends ServiceImpl<KongFuMapper, KongFu> impleme
     @Override
     public void mulSave(int n) {
         KongFu kongfu = getOne(new LambdaQueryWrapper<KongFu>().orderByDesc(KongFu::getId).last(" LIMIT 1"));
-        int start = kongfu == null ? 1 : kongfu.getId() + 1;
-        List<KongFu> kongfuList = IntStream.range(start, start+n).mapToObj(id -> {
+        long start = kongfu == null ? 1 : kongfu.getId() + 1;
+        List<KongFu> kongfuList = LongStream.range(start, start+n).mapToObj(id -> {
             KongFu kf = new KongFu();
             kf.setId(id);
             kf.setName("功夫" + id);
-            kf.setLevel(20 + id);
+            kf.setLevel((int)(20 + id));
             kf.setInventionUserId(id);
             return kf;
         }).collect(Collectors.toList());
